Recalculate `brct`
The `brct` variable might've gotten overwritten while iterating over multiple
alternate alleles on lines 195-206. If during the last iteration a brct entry
for the last alt was not found, it would set the `bcrt` variable to nil, which
would cause an error when trying to access it again on line 215. By
recalculating the `brct` variable again we ensure that the it will be set at
this point.